+++ projects/java_se7_sql/approaches/mybatis/spring/annot/src/main/java/com/meriosol/etr/dao/impl/EventDaoImpl.java
@@ -25,7 +25,7 @@ import java.util.Map;
 @Service("EventDaoService")
 public class EventDaoImpl implements EventDao {
     private static final Logger LOG = LoggerFactory.getLogger(EventDaoImpl.class);
-    private static final Long DEFAULT_MAX_EVENT_COUNT = 10000L; // normally client should query max 100
+    private static final Integer  DEFAULT_MAX_EVENT_COUNT = 10000; // normally client should query max 100
 
     @Autowired
     private EventMapper eventMapper;
@@ -89,7 +89,7 @@ public class EventDaoImpl implements EventDao {
      */
     @Override
     @Transactional
-    public List<Event> retrieveRecentEvents(Long maxEventCount) {
+    public List<Event> retrieveRecentEvents(Integer maxEventCount) {
         final String module = "retrieveRecentEvents";
         if (maxEventCount == null) {
             maxEventCount = DEFAULT_MAX_EVENT_COUNT;
